test: Support pytest-style fixtures
- Move factory-style Fixtures into Factories (create_X helpers)
- Add default_X fixtures for pytest
- Add session fixture for pytest
- Move sentry_sdk cleanup to pytest helper
- Move several database TestCase tests to standard unit tests (didnt require a database)
