+++ src/sentry/utils/pytest/__init__.py
@@ -3,4 +3,6 @@ from __future__ import absolute_import
 pytest_plugins = [
     'sentry.utils.pytest.sentry',
     'sentry.utils.pytest.selenium',
+    'sentry.utils.pytest.fixtures',
+    'sentry.utils.pytest.unittest',
 ]
